Update Filter to component/sub-component structure
