+++ src/components/Filter/Rule/EgoRule.js
@@ -3,12 +3,12 @@ import { connect } from 'react-redux';
 import { compose } from 'recompose';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { toPairs, has, includes, keys, flow, map, fromPairs } from 'lodash';
+import { toPairs, includes, keys } from 'lodash';
 import { SortableElement } from 'react-sortable-hoc';
 import DragHandle from './DragHandle';
 import DropDown from './DropDown';
 import Input from './Input';
-import { getVariableRegistry } from '../../selectors/protocol';
+import { getVariableRegistry } from '../../../selectors/protocol';
 
 const operators = toPairs({
   EXACTLY: 'is Exactly',
@@ -21,7 +21,7 @@ const operators = toPairs({
   LESS_THAN_OR_EQUAL: 'is Less Than or Exactly',
 });
 
-class EdgeRule extends PureComponent {
+class EgoRule extends PureComponent {
   static propTypes = {
     id: PropTypes.oneOfType([
       PropTypes.string,
@@ -30,33 +30,25 @@ class EdgeRule extends PureComponent {
     onUpdateRule: PropTypes.func,
     onDeleteRule: PropTypes.func,
     options: PropTypes.shape({
-      type: PropTypes.string,
-      operator: PropTypes.string,
       attribute: PropTypes.string,
+      operator: PropTypes.string,
       value: PropTypes.string,
     }),
-    edgeTypes: PropTypes.array,
-    edgeAttributes: PropTypes.object,
+    nodeAttributes: PropTypes.array,
     className: PropTypes.string,
   };
 
   static defaultProps = {
     options: {
-      type: '',
-      operator: '',
       attribute: '',
+      operator: '',
       value: '',
     },
-    edgeTypes: [],
-    edgeAttributes: {},
     onUpdateRule: () => {},
     onDeleteRule: () => {},
+    nodeAttributes: [],
     className: '',
   };
-
-  showAttributes() {
-    return has(this.props.edgeAttributes, this.props.options.type);
-  }
 
   showOperator() {
     return !!this.props.options.attribute;
@@ -70,36 +62,25 @@ class EdgeRule extends PureComponent {
   render() {
     const {
       id,
-      edgeTypes,
-      edgeAttributes,
+      nodeAttributes,
       onUpdateRule,
       onDeleteRule,
-      options: { type, operator, attribute, value },
+      options: { operator, attribute, value },
       className,
     } = this.props;
 
     return (
-      <div className={cx('rule', 'rule--edge', className)}>
+      <div className={cx('rule', 'rule--ego', className)}>
         <DragHandle />
         <div className="rule__options">
-          <div className="rule__option rule__option--type">
+          <div className="rule__option rule__option--attribute">
             <DropDown
-              options={edgeTypes}
-              value={type}
-              placeholder="{type}"
-              onChange={newValue => onUpdateRule(newValue, id, 'type')}
+              options={nodeAttributes}
+              value={attribute}
+              placeholder="{variable}"
+              onChange={newValue => onUpdateRule(newValue, id, 'attribute')}
             />
           </div>
-          {this.showAttributes() && (
-            <div className="rule__option rule__option--attribute">
-              <DropDown
-                options={has(edgeAttributes, type) ? edgeAttributes[type] : []}
-                value={attribute}
-                placeholder="{variable}"
-                onChange={newValue => onUpdateRule(newValue, id, 'attribute')}
-              />
-            </div>
-          )}
           { this.showOperator() && (
             <div className="rule__option rule__option--operator">
               <DropDown
@@ -110,7 +91,7 @@ class EdgeRule extends PureComponent {
               />
             </div>
           )}
-          { this.showValue() && (
+          {this.showValue() && (
             <div className="rule__option rule__option--value">
               <Input
                 value={value}
@@ -128,25 +109,14 @@ class EdgeRule extends PureComponent {
 function mapStateToProps(state) {
   const variableRegistry = getVariableRegistry(state);
 
-  const edgeAttributes = flow(
-    toPairs,
-    edgeTypes => map(
-      edgeTypes,
-      ([edgeType, options]) => [edgeType, keys(options.variables)],
-    ),
-    fromPairs,
-  );
-
   return {
-    edgeTypes: keys(variableRegistry.edge),
-    edgeAttributes: edgeAttributes(variableRegistry.edge),
+    nodeAttributes: keys(variableRegistry.node.person.variables),
   };
 }
 
-export { EdgeRule };
+export { EgoRule };
 
 export default compose(
   SortableElement,
   connect(mapStateToProps),
-)(EdgeRule);
-
+)(EgoRule);
