+++ src/components/index.js
@@ -1,3 +1,2 @@
 export { default as Timeline } from './Timeline';
-export { default as Rules } from './Rules';
 export { default as StaticImage } from './StaticImage';
